badrexpense/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.4
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
debugpy==1.7.0
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==3.2.25
django-environ==0.11.2
django-heroku==0.3.1
django-jazzmin==2.6.2
djangorestframework==3.15.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
gunicorn==22.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.7
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
jupyter_client==7.4.9
jupyter_core==4.12.0
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.21.6
openai==0.28.0
openpyxl==3.1.3
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.4.8
platformdirs==4.0.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
regex==2024.4.16
replicate==0.7.0
reportlab==4.2.2
requests==2.31.0
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tokenizers==0.13.3
torch==1.13.1
tornado==6.2
tqdm==4.66.5
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
validate-email==1.3
virtualenv==20.26.3
virtualenv-clone==0.5.7
wcwidth==0.2.13
whitenoise==6.5.0
yarl==1.9.4
zipp==3.15.0
